recover something that not important

+++ plugin/funwork/fortune.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/FloatTech/floatbox/web"
 	"github.com/FloatTech/zbputils/ctxext"
+	"github.com/tidwall/gjson"
 	zero "github.com/wdvxdr1123/ZeroBot"
 	"github.com/wdvxdr1123/ZeroBot/message"
 	"github.com/wdvxdr1123/ZeroBot/utils/helper"
@@ -139,18 +140,18 @@ func init() {
 				ctx.SendChain(message.At(user), message.Text(" 今天已经测过了哦~今日的人品值为", result[user], "呢~"))
 			}
 			// special time !
-			/*	if result[user] >= 90 && result[user] < 100 && egg[si] == 0 {
+			if result[user] >= 90 && result[user] < 100 && egg[si] == 0 {
 				egg[si] = (1)
-				img, err := web.RequestDataWith(web.NewDefaultClient(), "http://iw233.fgimax2.fgnwctvip.com/API/Ghs.php?type=json", "GET", Referer, ua)
+				img, err := web.GetData("https://api.lolicon.app/setu/v2?r18=1&num=1")
 				if err != nil {
 					ctx.SendChain(message.Text("ERROR:", err))
 					return
 				}
-				picURL := gjson.Get(string(img), "pic.0").String()
-				time.Sleep(time.Second * 2)
+				picURL := gjson.Get(string(img), "data.0.urls.original").String()
+				time.Sleep(time.Second * 3)
 				deleteme := ctx.SendChain(message.At(user), message.Text("\n这是今日奖励哦"), message.Text(picURL))
 				time.Sleep(time.Second * 20)
 				ctx.DeleteMessage(deleteme)
-			} */
+			}
 		})
 }
